@@ -1,11 +1,2 @@
-/// <reference path="node_modules/typescript/lib/lib.es2015.core.d.ts" />
-/// <reference path="node_modules/typescript/lib/lib.es2015.generator.d.ts" />
-/// <reference path="node_modules/typescript/lib/lib.es2015.proxy.d.ts" />
-/// <reference path="node_modules/typescript/lib/lib.es2015.reflect.d.ts" />
-/// <reference path="node_modules/typescript/lib/lib.es5.d.ts" />
-
-/// <reference path="node_modules/tns-core-modules-declarations/tns-core-modules.d.ts" />
-/// <reference path="node_modules/tns-core-modules-declarations/lib.dom.d.ts" />
-
-/// <reference path="node_modules/tns-platform-declarations/android/android17.d.ts" />
-/// <reference path="node_modules/tns-platform-declarations/ios/ios.d.ts" />
+/// <reference path="node_modules/tns-platform-declarations/android.d.ts" />
+/// <reference path="node_modules/tns-platform-declarations/ios.d.ts" />
